+++ source/source_estate/module_charge/chgmixing.cpp
@@ -1,6 +1,6 @@
 #include "source_estate/module_charge/chgmixing.h"
 #include "source_estate/update_pot.h"
-#include "source_lcao/module_dftu/dftu.h"
+#include "source_lcao/module_dftu/dftu_lcao.h"
 #include "source_lcao/module_deltaspin/spin_constrain.h"
 
 void module_charge::chgmixing_ks(const int iter, // scf iteration number
