+++ source/source_io/module_hs/write_vxc_r.hpp
@@ -2,7 +2,7 @@
 #define __WRITE_VXC_R_H_
 #include "source_io/module_parameter/parameter.h"
 #include "source_io/module_hs/write_hs_sparse.h"
-#include "source_lcao/module_dftu/op_dftu_lcao.h"
+#include "source_lcao/module_dftu/dftu_lcao_op_legacy.h"
 #include "source_lcao/module_operator_lcao/veff_lcao.h"
 #include "source_lcao/spar_hsr.h"
 #ifdef __EXX
