+++ source/source_lcao/lcao_set.h
@@ -12,7 +12,7 @@
 #include "source_basis/module_pw/pw_basis.h"
 #include "source_hamilt/module_surchem/surchem.h"
 #include "source_pw/module_pwdft/vl_pw.h"
-#include "source_lcao/module_dftu/dftu.h"
+#include "source_lcao/module_dftu/dftu_lcao.h"
 #include "source_lcao/setup_exx.h"
 #include "source_lcao/setup_deepks.h"
