+++ source/source_lcao/module_dftu/dftu_lcao_op_legacy.cpp
@@ -1,7 +1,7 @@
-#include "op_dftu_lcao.h"
+#include "dftu_lcao_op_legacy.h"
 #include "source_base/timer.h"
 #include "source_base/tool_title.h"
-#include "source_lcao/module_dftu/dftu.h"
+#include "source_lcao/module_dftu/dftu_lcao.h"
 
 namespace hamilt
 {
