+++ source/source_lcao/module_dftu/dftu_lcao_op_legacy.h
@@ -3,7 +3,7 @@
 
 #include "source_base/timer.h"
 #include "source_lcao/module_operator_lcao/operator_lcao.h"
-#include "source_lcao/module_dftu/dftu.h" // mohan add 20251107
+#include "source_lcao/module_dftu/dftu_lcao.h" // mohan add 20251107
 
 namespace hamilt
 {
