+++ source/source_lcao/module_dftu/dftu_tools.cpp
@@ -1,4 +1,4 @@
-#include "dftu.h"
+#include "dftu_lcao.h"
 #include "source_base/timer.h"
 #include "source_io/module_parameter/parameter.h"
