+++ source/source_lcao/module_dftu/dftu_yukawa.cpp
@@ -2,7 +2,7 @@
 #include "source_io/module_parameter/parameter.h"
 #include "source_base/constants.h"
 #include "source_base/global_function.h"
-#include "dftu.h"
+#include "dftu_lcao.h"
 
 #include <cmath>
 #include <complex>
