+++ source/source_lcao/module_dftu/test/dftu_lcao_test.cpp
@@ -5,8 +5,8 @@
 #define private public
 #include "source_io/module_parameter/parameter.h"
 #undef private
-#include "../dftu_lcao.h"
-#include "source_lcao/module_dftu/dftu.h"
+#include "../dftu_lcao_op.h"
+#include "source_lcao/module_dftu/dftu_lcao.h"
 
 Plus_U::Plus_U(){};
 Plus_U::~Plus_U(){};
