+++ source/source_pw/module_pwdft/force_pw.h
@@ -12,7 +12,7 @@
 #include "source_base/kernels/math_kernel_op.h"
 #include "source_psi/psi.h"
 #include "stru_fac.h"
-#include "source_lcao/module_dftu/dftu.h" // mohan add 2025-11-06
+#include "source_lcao/module_dftu/dftu_lcao.h" // mohan add 2025-11-06
 
 class pseudopot_cell_vnl;
