+++ source/source_pw/module_pwdft/force_pw_onsite.cpp
@@ -5,7 +5,7 @@
 #include "source_pw/module_pwdft/onsite_proj.h"
 #include "source_pw/module_pwdft/kernels/force_op.h"
 #include "source_io/module_parameter/parameter.h"
-#include "source_lcao/module_dftu/dftu.h"
+#include "source_lcao/module_dftu/dftu_lcao.h"
 #include "source_lcao/module_deltaspin/spin_constrain.h"
 
 template <typename FPTYPE, typename Device>
