+++ source/source_pw/module_pwdft/onsite_proj.cpp
@@ -6,7 +6,7 @@
 #include <tuple>
 #include "source_pw/module_pwdft/onsite_proj.h"
 #include "source_pw/module_pwdft/onsite_proj_print.h"
-#include "source_lcao/module_dftu/dftu.h"
+#include "source_lcao/module_dftu/dftu_lcao.h"
 #include "source_lcao/module_deltaspin/spin_constrain.h"
 #include "source_cell/cell_tools.h"
 #include "source_io/module_parameter/parameter.h"
