+++ source/source_pw/module_pwdft/op_pw_proj.h
@@ -5,7 +5,7 @@
 
 #include "source_cell/unitcell.h"
 #include "source_base/kernels/math_kernel_op.h"
-#include "source_lcao/module_dftu/dftu.h" // mohan add 20251106
+#include "source_lcao/module_dftu/dftu_lcao.h" // mohan add 20251106
 
 namespace hamilt {
