+++ source/source_pw/module_pwdft/stress_onsite.cpp
@@ -3,7 +3,7 @@
 #include "source_base/timer.h"
 #include "source_pw/module_pwdft/onsite_proj.h"
 #include "source_io/module_parameter/parameter.h"
-#include "source_lcao/module_dftu/dftu.h"
+#include "source_lcao/module_dftu/dftu_lcao.h"
 #include "source_lcao/module_deltaspin/spin_constrain.h"
 #include "stress_func.h"
